+++ checkout-ui-custom/src/_js/_v.custom.checkout.ui.js
@@ -291,7 +291,8 @@ class checkoutCustom {
       "span.shipping-date",
       ".shp-option-text-time",
       ".pkpmodal-pickup-point-sla",
-      ".shp-option-text-package"
+      ".shp-option-text-package",
+      ".srp-delivery-current-many__sla"
     ];
     try {
       $(`
@@ -299,7 +300,8 @@ class checkoutCustom {
         .vtex-omnishipping-1-x-leanShippingOption, 
         .vtex-omnishipping-1-x-packageItem:not(.v-changeShippingTimeInfo-active),
         .orderform-template .cart-template.mini-cart .item,
-        .vtex-pickup-points-modal-3-x-pickupPointSlaAvailability        
+        .vtex-pickup-points-modal-3-x-pickupPointSlaAvailability,
+        .srp-delivery-current-many
       `).each(function(i) {
         let txtselectin = $(this).find(mainSTIelems.map(elem => elem+":not(.v-changeShippingTimeInfo-elem-active)").join(", ")).text();
         if(txtselectin!="" && txtselectin.match(/(day)|(dia)/gm)) {
@@ -312,6 +314,28 @@ class checkoutCustom {
         }
         $(this).addClass("v-changeShippingTimeInfo-active");
       });
+
+      //temporaly
+      let shippingPreviewPackges = $(".srp-delivery-info .srp-packages:not(.v-changeShippingTimeInfo-elem-active)");
+      $(".js-shippingPreviewPackges").remove();
+      if(shippingPreviewPackges.length) {
+        var a = shippingPreviewPackges.text().split(":")[1].split(/,| and | e /);
+        var deliveryDates = [];
+        $.each(a, function(i) {
+            let txtselectin = a[i];
+            if(txtselectin!="" && txtselectin.match(/(day)|(dia)/gm)) {
+              let days = parseInt(txtselectin.match(/\d+/));
+              if(days) {
+                let _delivtext = _this.lang.deliveryDateText;
+                if(!! txtselectin.toLowerCase().match(/(ready in up)|(pronto)/gm)) _delivtext = _this.lang.PickupDateText; // check if is pickup. OBS: none of others solutions worked, needs constantly update
+                deliveryDates.push(`${_delivtext} <strong>${_this.addBusinessDays(days)}</strong>`);
+              }
+            }
+        })
+        shippingPreviewPackges.hide().after(`<p class="black-50 mt3 mb0 js-shippingPreviewPackges">${shippingPreviewPackges.text().split(":")[0]}: ${deliveryDates.join("; ")}</p>`).addClass("v-changeShippingTimeInfo-active");
+      } else {
+
+      }
     } catch(e) {}
   }
 
@@ -579,7 +603,6 @@ class checkoutCustom {
 
   init() {
     let _this = this;
-    
     _this.orderForm = vtexjs.checkout.orderForm ? vtexjs.checkout.orderForm : false;
     _this.general();
     _this.updateStep();
@@ -648,4 +671,3 @@ class checkoutCustom {
 }
 
 module.exports = checkoutCustom;
-